// app/src/models/entities/store-detail-module.ts
export interface StoreDetailModule {
    id: string;
    organization: string;
    organization_id: string;
    title: string;
    description_short: string;
    is_public: boolean;
    current_version: string;
    datenfeld_id: string;
    description: string;
    recent_changes: string;
    versions: string[];
}